feat: restructure functions and add commit function
