+++ deployer/dev/task/start_new_release.php
@@ -28,6 +28,5 @@ function setVersion(string $version): void {
 
 function commitVersion(string $version, string $type = VERSION_TYPE_BUILD): void {
     info("commit new version: \"{$version}\"");
-    runLocally("git add " . get('composer_path_app') . "/composer.json " . get('composer_path_app') . "/composer.lock");
-    runLocally("git commit " . get('dev_git_commit_options') . " -m \"" . get('dev_git_message_new_version_' . $type) . " " . $version . "\"");
+    commit(get('dev_git_message_new_version_' . $type) . " " . $version, [get('composer_path_app') . "/composer.json", get('composer_path_app') . "/composer.lock"]);
 }
